Add feedbacks ro admin namespace
