+++ app/helpers/admin/feedbacks_helper.rb
@@ -1,6 +1,6 @@
-module FeedbacksHelper
+module Admin::FeedbacksHelper
   def delete_feedback_link(feedback)
-    link_to 'Delete', feedback_path(feedback),
+    link_to 'Delete', admin_feedback_path(feedback),
             method: :delete,
             data: { confirm: 'Feedback will be deleted ?' }
   end
