+++ config/routes.rb
@@ -36,8 +36,12 @@ Rails.application.routes.draw do
 
   resources :companies
   resources :units
-  resources :feedbacks, only: %i[index show create destroy]
+  resources :feedbacks, only: :create
   resources :users
+
+  namespace :admin do
+    resources :feedbacks, only: %i[index show destroy]
+  end
 
   # Using :match so that error pages work for all types of requests, not just GET.
   match '/404', to: 'errors#not_found', via: :all
